container-management: Refactored tests imports
