+++ webserver/tests/fixtures/dockerhub_cr_fixtures.py
@@ -9,18 +9,18 @@ from app.models.registry import Registry
 DOCKER_CLASS = 'app.models.registry.DockerRegistry'
 
 @pytest.fixture
-def dockerhub_cr_name():
+def cr_name():
     return "acr.dockerhubcr.io"
 
 @pytest.fixture
-def dh_registry_client(mocker):
+def registry_client(mocker):
     mocker.patch(
         DOCKER_CLASS,
         return_value=Mock()
     )
 
 @pytest.fixture
-def dockerhub_cr_client(mocker):
+def cr_client(mocker):
     return mocker.patch(
         'app.helpers.container_registries.DockerRegistry',
         return_value=Mock(
@@ -30,7 +30,7 @@ def dockerhub_cr_client(mocker):
     )
 
 @pytest.fixture
-def dockerhub_cr_client_404(mocker):
+def cr_client_404(mocker):
     mocker.patch(
         DOCKER_CLASS,
         return_value=Mock(
@@ -40,18 +40,18 @@ def dockerhub_cr_client_404(mocker):
     )
 
 @pytest.fixture
-def dh_cr_class(mocker, dockerhub_cr_name, ):
-    return DockerRegistry(dockerhub_cr_name, creds={"user": "", "token": ""})
+def cr_class(mocker, cr_name, ):
+    return DockerRegistry(cr_name, creds={"user": "", "token": ""})
 
 @pytest.fixture
-def registry_dh(client, k8s_client, dockerhub_cr_name) -> Registry:
-    reg = Registry(dockerhub_cr_name, '', '')
+def registry(client, k8s_client, cr_name) -> Registry:
+    reg = Registry(cr_name, '', '')
     reg.add()
     return reg
 
 @pytest.fixture
-def container_dh(client, k8s_client, registry_dh, image_name) -> Container:
+def container(client, k8s_client, registry, image_name) -> Container:
     img, tag = image_name.split(':')
-    cont = Container(img, registry_dh, tag, True)
+    cont = Container(img, registry, tag, True)
     cont.add()
     return cont
